refactor: extract sequence routes into sequence_bp Blueprint
Move 5 sequence routes (/addParticipant, /addMessage,
/checkIfInsideParticipant, /getParticipantName, /editParticipantName)
into sequence/routes.py with a dedicated sequence_bp Blueprint.
Add None-guards to satisfy mypy in participant.py callers.
