+++ src/plantuml_gui/sequence/participant.py
@@ -49,8 +49,11 @@ def add_participant(puml: str, svg: str, clicked_x: int) -> str:
 
 def find_closest_participant(
     participants: List[Participant], target_cx: int
-) -> Participant:
+) -> Participant | None:
     """Find the participant with the closest cx value to the target_cx."""
+    if not participants:
+        return None
+
     min_distance = float("inf")
     closest_participant = participants[0]
 
@@ -87,6 +90,8 @@ def add_message(
     diagram = Diagram.from_svg(svg, puml)
     sender = find_closest_participant(diagram.participants, first_x)
     reciever = find_closest_participant(diagram.participants, second_x)
+    assert sender is not None
+    assert reciever is not None
 
     lines = puml.splitlines()
     lines.insert(-1, f"{sender.name} -> {reciever.name}: {message}")
@@ -96,6 +101,7 @@ def add_message(
 def get_participant_name(puml: str, svg: str, clicked_x: int):
     diagram = Diagram.from_svg(svg, puml)
     participant = find_closest_participant(diagram.participants, clicked_x)
+    assert participant is not None
 
     return participant.name
 
@@ -103,5 +109,6 @@ def get_participant_name(puml: str, svg: str, clicked_x: int):
 def edit_participant_name(puml: str, svg: str, newname: str, clicked_x: int):
     diagram = Diagram.from_svg(svg, puml)
     participant = find_closest_participant(diagram.participants, clicked_x)
+    assert participant is not None
 
     return puml.replace(participant.name, newname)
